+++ event_reporter/reporter/application_errors_parser.go
@@ -92,7 +92,24 @@ func parseResourceSyncResultErrors(rs *appv1.ResourceStatus, os *appv1.Operation
 	return errors
 }
 
-func parseAggregativeHealthErrors(rs *appv1.ResourceStatus, apptree *appv1.ApplicationTree) []*events.ObjectError {
+func parseAggregativeHealthErrorsOfApplication(a *appv1.Application, appTree *appv1.ApplicationTree) []*events.ObjectError {
+	var errors []*events.ObjectError
+	if a.Status.Resources == nil {
+		return errors
+	}
+
+	for _, rs := range a.Status.Resources {
+		if rs.Health != nil {
+			if rs.Health.Status != health.HealthStatusHealthy {
+				errors = append(errors, parseAggregativeHealthErrors(&rs, appTree, true)...)
+			}
+		}
+	}
+
+	return errors
+}
+
+func parseAggregativeHealthErrors(rs *appv1.ResourceStatus, apptree *appv1.ApplicationTree, addReference bool) []*events.ObjectError {
 	errs := make([]*events.ObjectError, 0)
 
 	if apptree == nil {
@@ -108,12 +125,24 @@ func parseAggregativeHealthErrors(rs *appv1.ResourceStatus, apptree *appv1.Appli
 
 	for _, cn := range childNodes {
 		if cn.Health != nil && cn.Health.Status == health.HealthStatusDegraded {
-			errs = append(errs, &events.ObjectError{
+			newErr := events.ObjectError{
 				Type:     "health",
 				Level:    "error",
 				Message:  cn.Health.Message,
 				LastSeen: *cn.CreatedAt,
-			})
+			}
+
+			if addReference {
+				newErr.SourceReference = events.ErrorSourceReference{
+					Group:     rs.Group,
+					Version:   rs.Version,
+					Kind:      rs.Kind,
+					Namespace: rs.Namespace,
+					Name:      rs.Name,
+				}
+			}
+
+			errs = append(errs, &newErr)
 		}
 	}